Refactoring Ballistic engine.
