+++ LawlerBallisticsDesk/Classes/Ballistics.cs
@@ -172,776 +172,9 @@ namespace LawlerBallisticsDesk.Classes
 
         #region "Public Properties"
         /// <summary>
-        /// The latitude of the target at zero in decimal degrees (DD).
-        /// </summary>
-        public double zTargetLat
-        {
-            get
-            {
-                if (_zTargetLatitude == 0) _zTargetLatitude = 34.681129;
-                return _zTargetLatitude;
-            }
-            set
-            {
-                _zTargetLatitude = value;
-                RaisePropertyChanged(nameof(zTargetLat));
-            }
-        }
-        /// <summary>
-        /// The longitude of the target at zero in decimal degrees (DD).
-        /// </summary>
-        public double zTargetLon
-        {
-            get
-            {
-                if (_zTargetLongitude == 0) _zTargetLongitude = -86.987622;
-                return _zTargetLongitude;
-            }
-            set
-            {
-                _zTargetLongitude = value;
-                RaisePropertyChanged(nameof(zTargetLon));
-            }
-        }
-        /// <summary>
-        /// The latitude of the shooter at zero time in decimal degrees (DD).
-        /// </summary>
-        public double zShooterLat
-        {
-            get
-            {
-                if (_zShooterLatitude == 0) _zShooterLatitude = 34.681129;
-                return _zShooterLatitude;
-            }
-            set
-            {
-                _zShooterLatitude = value;
-                RaisePropertyChanged(nameof(zShooterLat));
-            }
-        }
-        /// <summary>
-        /// The longitude of the shooter at zero in decimal degrees (DD).
-        /// </summary>
-        public double zShooterLon
-        {
-            get
-            {
-                if (_zShooterLongitude == 0) _zShooterLongitude = -86.988711;
-                return _zShooterLongitude;
-            }
-            set
-            {
-                _zShooterLongitude = value;
-                RaisePropertyChanged(nameof(zShooterLon));
-            }
-        }
-        /// <summary>
-        /// The longitude of the shooter in decimal degrees (DD).
-        /// </summary>
-        public double ShooterLon
-        {
-            get
-            {
-                if (_ShooterLongitude == 0) _ShooterLongitude = -86.988711;
-                return _ShooterLongitude;
-            }
-            set
-            {
-                _ShooterLongitude = value;
-                RaisePropertyChanged(nameof(ShooterLon));
-            }
-        }
-        /// <summary>
-        /// The latitude of the shooter in decimal degrees (DD).
-        /// </summary>
-        public double ShooterLat
-        {
-            get
-            {
-                if (_ShooterLatitude == 0) _ShooterLatitude = 34.681129;
-                return _ShooterLatitude;
-            }
-            set
-            {
-                _ShooterLatitude = value;
-                RaisePropertyChanged(nameof(ShooterLat));
-            }
-        }
-        /// <summary>
-        /// The line of sight distance to the target in yards.
-        /// </summary>
-        public double ShotDistance { get { return _ShotDistance; } set { _ShotDistance = value; RaisePropertyChanged(nameof(ShotDistance)); } }
-        /// <summary>
-        /// The absolute value of the vertical angle of the shot.
-        /// </summary>
-        public double ShotAngle { get { return _ShotAngle; } set { _ShotAngle = value; RaisePropertyChanged(nameof(ShotAngle)); } }
-        /// <summary>
-        /// The latitude of the target in decimal degrees (DD).
-        /// </summary>
-        public double TargetLat
-        {
-            get
-            {
-                if (_TargetLatitude == 0) _TargetLatitude = 34.681129;
-                return _TargetLatitude;
-            }
-            set
-            {
-                _TargetLatitude = value;
-                RaisePropertyChanged(nameof(TargetLat));
-            }
-        }
-        /// <summary>
-        /// The longitude of the target at decimal degrees (DD).
-        /// </summary>
-        public double TargetLon
-        {
-            get
-            {
-                if (_TargetLongitude == 0) _TargetLongitude = -86.987622;
-                return _TargetLongitude;
-            }
-            set
-            {
-                _TargetLongitude = value;
-                RaisePropertyChanged(nameof(TargetLon));
-            }
-        }
-        /// <summary>
-        /// The current relative humidity.
-        /// </summary>
-        public double RelHumidity
-        {
-            get
-            {
-                if (_RelHumidity == 0) _RelHumidity = 50.00;
-                return _RelHumidity;
-            }
-            set
-            {
-                _RelHumidity = value;
-                RaisePropertyChanged(nameof(RelHumidity));
-            }
-        }
-        /// <summary>
-        /// The relative humidity when zeroed.
-        /// </summary>
-        public double zRelHumidity
-        {
-            get
-            {
-                if (_zRelHumidity == 0) _zRelHumidity = 50.00;
-                return _zRelHumidity;
-            }
-            set
-            {
-                _zRelHumidity = value;
-                RaisePropertyChanged(nameof(zRelHumidity));
-            }
-        }
-        /// <summary>
-        /// The barometric pressure in inches Hg.
-        /// </summary>
-        public double BaroPressure
-        {
-            get
-            {
-                if (_BaroPressure == 0) _BaroPressure = 29.92;
-                return _BaroPressure;
-            }
-            set
-            {
-                _BaroPressure = value;
-                RaisePropertyChanged(nameof(BaroPressure));
-            }
-        }
-        /// <summary>
-        /// The barometric pressure in inches Hg at sight in.
-        /// </summary>
-        public double zBaroPressure
-        {
-            get
-            {
-                if (_zBaroPressure == 0) _zBaroPressure = 29.92;
-                return _zBaroPressure;
-            }
-            set
-            {
-                _zBaroPressure = value;
-                RaisePropertyChanged(nameof(zBaroPressure));
-            }
-        }
-        /// <summary>
-        /// Density Altitude
-        /// </summary>
-        public double DensityAlt
-        {
-            get
-            {
-                return _DensityAlt;
-            }
-            set {
-                _DensityAlt = value;
-                RaisePropertyChanged(nameof(DensityAlt));
-            }
-        }
-        /// <summary>
-        /// Density Altitude at time of zero.
-        /// </summary>
-        public double zDensityAlt
-        {
-            get
-            {
-                return _zDensityAlt;
-            }
-            set
-            {
-                _zDensityAlt = value;
-                RaisePropertyChanged(nameof(zDensityAlt));
-            }
-        }
-        /// <summary>
-        /// Ambient Temperature in fahrenheit.
-        /// </summary>
-        public double TempF
-        {
-            get
-            {
-                if (_TempF == 0) _TempF = 60.00;
-                return _TempF;
-            }
-            set
-            {
-                _TempF = value;
-                RaisePropertyChanged(nameof(TempF));
-            }
-        }
-        /// <summary>
-        /// Ambient Temperature in fahrenheit when zeroed.
-        /// </summary>
-        public double zTempF
-        {
-            get
-            {
-                if (_zTempF == 0) _zTempF = 60.00;
-                return _zTempF;
-            }
-            set
-            {
-                _zTempF = value;
-                RaisePropertyChanged(nameof(zTempF));
-            }
-        }
-        /// <summary>
-        /// Measured from centerline of scope to centerline of barrel.
-        /// </summary>
-        public double ScopeHeight
-        {
-            get
-            {
-                if (_ScopeHeight == 0) _ScopeHeight = 2.50;
-                return _ScopeHeight;
-            }
-            set
-            {
-                _ScopeHeight = value;
-                RaisePropertyChanged(nameof(ScopeHeight));
-                RaisePropertyChanged(nameof(Fo));
-                RaisePropertyChanged(nameof(MidRange));
-                RaisePropertyChanged(nameof(ZeroMaxRise));
-                RaisePropertyChanged(nameof(PointBlankRange));
-                RaisePropertyChanged(nameof(ZeroNearRange));
-            }
-        }
-        /// <summary>
         /// Used to estimate a bullet's BC and for initial Zone scale multiplier value selection.
         /// </summary>
         public BulletShapeEnum BulletShapeType { get { return _BulletShapeType; } }
-        public double MuzzleVelocity
-        {
-            get
-            {
-                return _MuzzleVelocity;
-            }
-            set
-            {
-                _MuzzleVelocity = value;
-                RaisePropertyChanged(nameof(MuzzleVelocity));
-                RaisePropertyChanged(nameof(Fo));
-                RaisePropertyChanged(nameof(MidRange));
-                RaisePropertyChanged(nameof(ZeroMaxRise));
-                RaisePropertyChanged(nameof(PointBlankRange));
-                RaisePropertyChanged(nameof(ZeroNearRange));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// Maximum achievable distance (yards)
-        /// </summary>
-        public double MaximumRange
-        {
-            get
-            {
-                return MaxRange();
-            }
-        }
-        /// <summary>
-        /// Fo = the initial drag/retard factor starting at the muzzle.  Calculated from V1, V2, and the
-        /// distance between where the two velocities were measured.  If V1 is not at the muzzle then
-        /// the muzzle velocity will be calculated.  Alternatively can be calculated from the bullet BC
-        /// and the measured muzzle velocity.
-        /// </summary>
-        public double Fo
-        {
-            get
-            {
-                return _Fo;
-            }
-            set
-            {
-                _Fo = value;
-                RaisePropertyChanged(nameof(Fo));
-                RaisePropertyChanged(nameof(MidRange));
-                RaisePropertyChanged(nameof(ZeroMaxRise));
-                RaisePropertyChanged(nameof(PointBlankRange));
-                RaisePropertyChanged(nameof(ZeroNearRange));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// F2 = the initial drag/retard factor starting at the begining of Zone 2.
-        /// </summary>
-        public double F2
-        {
-            get
-            {
-                //F = BC(cV^0.5)
-                if (_F2 == 0) _F2 = BCz2 * (166 * Math.Pow(Zone2TransSpeed, 0.5));
-                return _F2;
-            }
-            set
-            {
-                _F2 = value;
-                RaisePropertyChanged(nameof(F2));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// F3 = the initial drag/retard factor starting at the begining if zone 3.
-        /// </summary>
-        public double F3
-        {
-            get
-            {
-                //F = (cV^0.5)
-                if (_F3 == 0) _F3 = (166 * Math.Pow(Zone3TransSpeed, 0.5));
-                return _F3;
-            }
-            set
-            {
-                _F3 = value;
-                RaisePropertyChanged(nameof(F3));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// F4 = the initial drag/retard factor starting at the end of zone 3.
-        /// </summary>
-        public double F4
-        {
-            get
-            {
-                //F = (cV^0.5)
-                if (_F4 == 0) _F4 = (166 * Math.Pow(Zone3TransSpeed, 0.5));
-                return _F4;
-            }
-            set
-            {
-                _F4 = value;
-                RaisePropertyChanged(nameof(F4));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// The first velocity measured.  This value should be taken as close to the muzzle as possible.
-        /// </summary>
-        public double V1 { get { return _V1; } set { _V1 = value; RaisePropertyChanged(nameof(V1)); } }
-        /// <summary>
-        /// The second velocity measured.  Preferably 300 feet or more from V1 measurement location.
-        /// </summary>
-        public double V2 { get { return _V2; } set { _V2 = value; RaisePropertyChanged(nameof(V2)); } }
-        /// <summary>
-        /// Distance in feet from muzzle where V1 is measured.
-        /// </summary>
-        public double D1 { get { return _D1; } set { _D1 = value; RaisePropertyChanged(nameof(D1)); } }
-        /// <summary>
-        /// Distance in feet from the muzzle to where the second velocity reading (V2) was recorded.
-        /// It is reccommeded to have atleast 300 ft between D1 and D2, 600 results in 1/2 the error
-        /// as 300, 800 ft is ideal.
-        /// </summary>
-        public double D2 { get { return _D2; } set { _D2 = value; RaisePropertyChanged(nameof(D2)); } }
-        /// <summary>
-        /// Ballistic Coefficient G1
-        /// </summary>
-        public double BCg1 { get { return _BCg1; } set { _BCg1 = value; RaisePropertyChanged(nameof(BCg1)); } }
-        public double BCz2 {
-            get
-            {
-                if (_BCz2 == 0) _BCz2 = BCg1;
-                return _BCz2;
-            }
-            set { _BCz2 = value; RaisePropertyChanged(nameof(BCz2)); } }
-        /// <summary>
-        /// The maximum bullet rise from the muzzle to the zero range.
-        /// </summary>
-        public double ZeroMaxRise
-        {
-            get
-            {
-                if ((UseMaxRise) & (_Hm == 0)) _Hm = 2.00;
-                return _Hm;
-            }
-            set 
-            {
-                _Hm = value;
-                RaisePropertyChanged(nameof(ZeroMaxRise));
-                RaisePropertyChanged(nameof(MidRange));
-                RaisePropertyChanged(nameof(PointBlankRange));
-                RaisePropertyChanged(nameof(ZeroNearRange));
-            }
-        }
-        /// <summary>
-        /// The range the gun is sighted in to have zero vertical deviation.
-        /// </summary>
-        public double ZeroRange
-        {
-            get
-            {
-                if ((_ZeroRange == 0)&(!UseMaxRise)) _ZeroRange = 200.00;
-                return _ZeroRange;
-            }
-            set
-            {
-                _ZeroRange = value;
-                RaisePropertyChanged(nameof(ZeroRange));
-            }
-        }
-        /// <summary>
-        /// True if the zero range is to be calculated from the maximum desired rise above the sight plane
-        /// on the trajectory to zero.
-        /// </summary>
-        public bool UseMaxRise { get { return _UseMaxRise; } set { _UseMaxRise = value; RaisePropertyChanged(nameof(UseMaxRise)); } }
-        /// <summary>
-        /// Midrange where PBR height (Hm) or maximum rise occurs
-        /// </summary>
-        public double MidRange
-        {
-            get
-            {
-                double lMR;
-                //M = 1/(((G/Vo)/((Hm+S)^0.5)) + 2/Fo)
-                lMR = 1 / (((_G / MuzzleVelocity) / (Math.Pow((_Hm + ScopeHeight), 0.5))) + 2 / Fo);
-                return lMR;
-            }
-        }
-        /// <summary>
-        /// Calculate Point Blank Range Distance
-        /// </summary>
-        public double PointBlankRange
-        {
-            get
-            {
-                double lSQ; double lP;
-
-                //P = (1 + SQ)/(1/Fo + SQ/M)
-                //SQ  = SH/2^0.5
-                lSQ = SH() / (Math.Pow(2, 0.5));
-                lP = (1 + lSQ) / ((1 / Fo) + (lSQ / MidRange));
-                return lP;
-            }
-        }
-        /// <summary>
-        /// The near range where the bullet crosses the sight plane on the way to the far zero vertical deviation.
-        /// </summary>
-        public double ZeroNearRange 
-        {
-            get
-            {
-                double lNZ;
-
-                //Zn = (1 - SH) / (1 / Fo - SH / M)
-                lNZ = (1 - SH()) / ((1 / Fo) - (SH()/ MidRange));
-
-                return lNZ;
-            }
-        }
-        /// <summary>
-        /// The speed where Supersonic enters first transonic zone approx 1.2 Mach or 1340 fps.
-        /// </summary>
-        public double Zone1TransSpeed
-        {
-            get
-            {
-                double lZ1;
-                double lDA;
-
-                Atmospherics lA = new Atmospherics();
-                lA.Temp = TempF;
-                lA.TempUnits = "F";
-                lA.HumidityRel = RelHumidity;
-                lA.Pressure = BaroPressure;
-                lA.PressureUnits = "inHg";
-                lZ1 = lA.SpeedOfSound;
-                lZ1 = lZ1 * Zone1MachFactor;
-                lDA = lA.DensityAlt;
-                _Zone1TransSpeed = lZ1;
-                if (_Zone1TransSpeed == 0) _Zone1TransSpeed = 1340;
-                return _Zone1TransSpeed;
-            }
-        }
-        /// <summary>
-        /// The speed where the first transonic zone enters the second transonic zone, approx 1.0 Mach or 1125 fps.
-        /// </summary>
-        public double Zone2TransSpeed
-        {
-            get
-            {
-                double lZ2;
-                Atmospherics lA = new Atmospherics();
-                lA.Temp = TempF;
-                lA.TempUnits = "F";
-                lA.HumidityRel = RelHumidity;
-                lA.Pressure = BaroPressure;
-                lA.PressureUnits = "inHg";
-                lZ2 = lA.SpeedOfSound;
-                lZ2 = lZ2 * Zone2MachFactor;
-                _Zone2TransSpeed = lZ2;
-                if (_Zone2TransSpeed == 0) _Zone2TransSpeed = 1125;
-                return _Zone2TransSpeed;
-            }
-        }
-        /// <summary>
-        /// The speed where the second transonic zone enters the subsonic zone, approx 890 fps.
-        /// </summary>
-        public double Zone3TransSpeed
-        {
-            get
-            {
-                double lZ3;
-                Atmospherics lA = new Atmospherics();
-                lA.Temp = TempF;
-                lA.TempUnits = "F";
-                lA.HumidityRel = RelHumidity;
-                lA.Pressure = BaroPressure;
-                lA.PressureUnits = "inHg";
-                lZ3 = lA.SpeedOfSound;
-                lZ3 = lZ3 * Zone3MachFactor;
-                _Zone3TransSpeed = lZ3;
-                if (_Zone3TransSpeed == 0) _Zone3TransSpeed = 890;
-                return _Zone3TransSpeed;
-            }          
-        }
-        /// <summary>
-        /// The factor of the speed of sound to transition drag zones.
-        /// </summary>
-        public double Zone1MachFactor 
-        {
-            get
-            {
-                if (_Zone1MachFactor == 0) _Zone1MachFactor = 1.2;
-                return _Zone1MachFactor;
-            }
-            set
-            {
-                _Zone1MachFactor = value;
-                RaisePropertyChanged(nameof(Zone1MachFactor));
-            }
-        }
-        /// <summary>
-        /// The factor of the speed of sound to transition drag zones.
-        /// </summary>
-        public double Zone2MachFactor
-        {
-            get
-            {
-                if (_Zone2MachFactor == 0) _Zone2MachFactor = 1;
-                return _Zone2MachFactor;
-            }
-            set
-            {
-                _Zone2MachFactor = value;
-                RaisePropertyChanged(nameof(Zone2MachFactor));
-            }
-        }
-        /// <summary>
-        /// The factor of the speed of sound to transition drag zones.
-        /// </summary>
-        public double Zone3MachFactor
-        {
-            get
-            {
-                if (_Zone3MachFactor == 0) _Zone3MachFactor = 0.8;
-                return _Zone3MachFactor;
-            }
-            set
-            {
-                _Zone3MachFactor = value;
-                RaisePropertyChanged(nameof(Zone3MachFactor));
-            }
-        }
-        /// <summary>
-        /// A slope adjustment term based on the bullet design.  Default is 0.78.
-        /// </summary>
-        public double Zone1SlopeMultiplier
-        {
-            //TODO: in the interface recommend a multiplier based on the selected bullet type.
-            //TODO  If this is zero used input the recommended value based on bullet type.
-            //TODO if the bullet type selection changes and the value is a default of another
-            // type then change it to the recommended value for the currently selected type.
-
-            get
-            {
-                if (_Zone1SlopeMultiplier == 0) _Zone1SlopeMultiplier = 0.75;
-                return _Zone1SlopeMultiplier;
-            }
-            set
-            {
-                _Zone1SlopeMultiplier = value;
-                RaisePropertyChanged(nameof(Zone1SlopeMultiplier));
-            }
-        }
-        /// <summary>
-        /// A slope adjustment term based on the bullet design.  Default is 0.78.
-        /// </summary>
-        public double Zone3SlopeMultiplier
-        {
-            get
-            {
-                if (_Zone3SlopeMultiplier == 0) _Zone3SlopeMultiplier = 0.75;
-                return _Zone3SlopeMultiplier;
-            }
-            set
-            {
-                _Zone3SlopeMultiplier = value;
-                RaisePropertyChanged(nameof(Zone3SlopeMultiplier));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// Slope value for Zone 1 (supersonic) also represented as N for zone 1.
-        /// </summary>
-        public double Zone1Slope
-        {
-            get
-            {
-                if (_Zone1Slope == 0) _Zone1Slope = 0.5;
-                return _Zone1Slope;
-            }
-            set
-            {
-                _Zone1Slope = value;
-                RaisePropertyChanged(nameof(Zone1Slope));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-            }
-        }
-        /// <summary>
-        /// A correction factor for drop resulting from the downward angle of the bullet as a result of zone 1 arc.
-        /// </summary>
-        public double Zone1AngleFactor {
-            get
-            {
-                if (_Zone1AngleFactor == 0) _Zone1AngleFactor = 0.6957;
-                return _Zone1AngleFactor;
-            }
-            set { _Zone1AngleFactor = value; RaisePropertyChanged(nameof(Zone1AngleFactor)); }
-        }
-        /// <summary>
-        /// Slope value for Zone 3 (Transonic) also represented as N for zone 3.
-        /// </summary>
-        public double Zone3Slope
-        {
-            get
-            {
-                if (_Zone3Slope == 0) _Zone3Slope = -4;
-                return _Zone3Slope;
-            }
-            set
-            {
-                _Zone3Slope = value;
-                RaisePropertyChanged(nameof(Zone3Slope));
-                RaisePropertyChanged(nameof(Zone1Range));
-                RaisePropertyChanged(nameof(Zone2Range));
-                RaisePropertyChanged(nameof(Zone3Range));
-
-            }
-        }
-        /// <summary>
-        /// The range where the bullet passes our of Zone 1 into zone 2.
-        /// </summary>
-        public double Zone1Range
-        {
-            get
-            {
-                //R = (2/3)*Fo(1 - (V/Vo)^N)
-                _Zone1Range = (2.00 / 3.00) * Fo * (1 - Math.Pow((Zone1TransSpeed / MuzzleVelocity), Zone1Slope));
-                if (_Zone1Range < 0) _Zone1Range = 0;
-                return _Zone1Range;
-            }           
-        }
-        /// <summary>
-        /// The range where the bullet passes our of Zone 2 into zone 3.
-        /// </summary>
-        public double Zone2Range
-        {
-            get
-            {
-                //R = F/(3ln(Vo/V))
-                if (MuzzleVelocity < Zone2TransSpeed)
-                {
-                    _Zone2Range = 0;
-                }
-                else
-                {
-                    _Zone2Range = (F2 / 3) * Math.Log(Zone1TransSpeed / Zone2TransSpeed);
-                    _Zone2Range = _Zone2Range + Zone1Range;
-                }
-                return _Zone2Range;
-            }           
-        }
-        /// <summary>
-        /// The range where the bullet passes our of Zone 3 into zone 4.
-        /// </summary>
-        public double Zone3Range
-        {
-            get
-            {
-                //R = (F2-F3)/3N
-                if (MuzzleVelocity < Zone3TransSpeed)
-                {
-                    _Zone3Range = 0;
-                }
-                else
-                {
-                    _Zone3Range = (F2 - F3) / (3 * Zone3Slope);
-                    _Zone3Range = _Zone3Range + Zone2Range;
-                }
-                return _Zone3Range;
-            }           
-        }
         public Barrel SelectedBarrel { get { return _SelectedBarrel; } }
         /// <summary>
         /// Number of inches of barrel length per rev of the rifling.
@@ -1921,7 +1154,6 @@ namespace LawlerBallisticsDesk.Classes
             }
 
         }
-
         /// <summary>
         /// Calculates the horizontal and vertical devation of the flight path due to the Coriolis effect.
         /// </summary>
@@ -2062,19 +1294,6 @@ namespace LawlerBallisticsDesk.Classes
                 lFa = F4;
             }
             return lFa;
-        }
-        /// <summary>
-        /// Factor used to calculate Zero range, Near Zero Range, and Point-Blank-Range (PBR)
-        /// </summary>
-        /// <returns></returns>
-        private double SH()
-        {
-            double lSH;
-
-            if (_Hm == 0) return 0;
-            //SH = (1 + S/Hm)^0.5
-            lSH = Math.Pow((1 + (ScopeHeight / _Hm)), 0.5);
-            return lSH;
         }
         /// <summary>
         /// Magnitude of the horizontal displacement induced by the bullets gyroscopic force.
